Partial cleanup of SpanExtensionsLab (#1348)
* Partial cleanup of SpanExtensionsLab

* Adding back IndexOf overloads for now

* Removing IndexOf overloads from buffers.experimental

* Reverting JsonParser change back to what it was originally.
