+++ src/System.Binary/System/Binary/BufferWriter.cs
@@ -27,5 +27,36 @@ namespace System.Binary
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public static void WriteLittleEndian<[Primitive]T>(this Span<byte> span, T value) where T : struct
             => span.Write(BitConverter.IsLittleEndian ? value : UnsafeUtilities.Reverse(value));
+
+
+
+        /// <summary>
+        /// Writes a structure of type T into a slice of bytes.
+        /// </summary>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static void Write<[Primitive]T>(this Span<byte> slice, T value)
+            where T : struct
+        {
+            if ((uint)Unsafe.SizeOf<T>() > (uint)slice.Length)
+            {
+                throw new ArgumentOutOfRangeException();
+            }
+            Unsafe.WriteUnaligned<T>(ref slice.DangerousGetPinnableReference(), value);
+        }
+
+        /// <summary>
+        /// Writes a structure of type T into a slice of bytes.
+        /// </summary>
+        [MethodImpl(MethodImplOptions.AggressiveInlining)]
+        public static bool TryWrite<[Primitive]T>(this Span<byte> slice, T value)
+            where T : struct
+        {
+            if (Unsafe.SizeOf<T>() > (uint)slice.Length)
+            {
+                return false;
+            }
+            Unsafe.WriteUnaligned<T>(ref slice.DangerousGetPinnableReference(), value);
+            return true;
+        }
     }
 }
