+++ src/System.Buffers.Primitives/System/SpanExtensions.cs
@@ -13,36 +13,6 @@ namespace System
     /// </summary>
     public static partial class SpanExtensionsLabs
     {
-        public static bool StartsWith(this ReadOnlySpan<byte> bytes, ReadOnlySpan<byte> slice)
-        {
-            if (slice.Length > bytes.Length)
-            {
-                return false;
-            }
-
-            for (int i = 0; i < slice.Length; i++)
-            {
-                if (bytes[i] != slice[i])
-                {
-                    return false;
-                }
-            }
-
-            return true;
-        }
-
-        public static bool StartsWith<T>(this ReadOnlySpan<T> items, ReadOnlySpan<T> slice)
-            where T : struct, IEquatable<T>
-        {
-            if (slice.Length > items.Length) return false;
-            return items.Slice(0, slice.Length).SequenceEqual(slice);
-        }
-
-        public static int SequentialIndexOf(this Span<byte> span, byte value)
-        {
-            return SequentialIndexOf(ref span.DangerousGetPinnableReference(), value, span.Length);
-        }
-
         public unsafe static int IndexOfAny(this Span<byte> buffer, byte value0, byte value1)
         {
             fixed (byte* pSearchSpace = &buffer.DangerousGetPinnableReference())
@@ -57,11 +27,6 @@ namespace System
             {
                 return IndexOfAny(pSearchSpace, buffer.Length, value0, value1, value2);
             }
-        }
-
-        public static int SequentialIndexOf(this ReadOnlySpan<byte> span, byte value)
-        {
-            return SequentialIndexOf(ref span.DangerousGetPinnableReference(), value, span.Length);
         }
 
         public unsafe static int IndexOfAny(this ReadOnlySpan<byte> buffer, byte value0, byte value1)
@@ -79,80 +44,7 @@ namespace System
                 return IndexOfAny(pSearchSpace, buffer.Length, value0, value1, value2);
             }
         }
-
-        private static unsafe int SequentialIndexOf(ref byte searchSpace, byte value, int length)
-        {
-            Debug.Assert(length >= 0);
-
-            IntPtr index = (IntPtr)0; // Use IntPtr for arithmetic to avoid unnecessary 64->32->64 truncations
-            while (length >= 8)
-            {
-                length -= 8;
-
-                if (value == Unsafe.Add(ref searchSpace, index))
-                    goto Found;
-                if (value == Unsafe.Add(ref searchSpace, index + 1))
-                    goto Found1;
-                if (value == Unsafe.Add(ref searchSpace, index + 2))
-                    goto Found2;
-                if (value == Unsafe.Add(ref searchSpace, index + 3))
-                    goto Found3;
-                if (value == Unsafe.Add(ref searchSpace, index + 4))
-                    goto Found4;
-                if (value == Unsafe.Add(ref searchSpace, index + 5))
-                    goto Found5;
-                if (value == Unsafe.Add(ref searchSpace, index + 6))
-                    goto Found6;
-                if (value == Unsafe.Add(ref searchSpace, index + 7))
-                    goto Found7;
-
-                index += 8;
-            }
-
-            if (length >= 4)
-            {
-                length -= 4;
-
-                if (value == Unsafe.Add(ref searchSpace, index))
-                    goto Found;
-                if (value == Unsafe.Add(ref searchSpace, index + 1))
-                    goto Found1;
-                if (value == Unsafe.Add(ref searchSpace, index + 2))
-                    goto Found2;
-                if (value == Unsafe.Add(ref searchSpace, index + 3))
-                    goto Found3;
-
-                index += 4;
-            }
-
-            while (length > 0)
-            {
-                if (value == Unsafe.Add(ref searchSpace, index))
-                    goto Found;
-
-                index += 1;
-                length--;
-            }
-            return -1;
-
-            Found: // Workaround for https://github.com/dotnet/coreclr/issues/9692
-            return (int)(byte*)index;
-            Found1:
-            return (int)(byte*)(index + 1);
-            Found2:
-            return (int)(byte*)(index + 2);
-            Found3:
-            return (int)(byte*)(index + 3);
-            Found4:
-            return (int)(byte*)(index + 4);
-            Found5:
-            return (int)(byte*)(index + 5);
-            Found6:
-            return (int)(byte*)(index + 6);
-            Found7:
-            return (int)(byte*)(index + 7);
-        }
-
+        
         private static unsafe int IndexOfAny(byte* searchSpace, int length, byte value0, byte value1)
         {
             var offset = 0;
@@ -302,120 +194,6 @@ namespace System
             return offset;
         }
 
-        public static bool TryIndicesOf(this Span<byte> buffer, byte value, Span<int> indices, out int numberOfIndices)
-        {
-            var length = buffer.Length;
-            if (length == 0 || indices.Length == 0)
-            {
-                numberOfIndices = 0;
-                return false;
-            }
-
-            return TryIndicesOf(ref buffer.DangerousGetPinnableReference(), value, length, indices, out numberOfIndices);
-        }
-
-        public static bool TryIndicesOf(this ReadOnlySpan<byte> buffer, byte value, Span<int> indices, out int numberOfIndices)
-        {
-            var length = buffer.Length;
-            if (length == 0 || indices.Length == 0)
-            {
-                numberOfIndices = 0;
-                return false;
-            }
-
-            return TryIndicesOf(ref buffer.DangerousGetPinnableReference(), value, length, indices, out numberOfIndices);
-        }
-
-        private unsafe static bool TryIndicesOf(ref byte searchSpace, byte value, int length, Span<int> indices, out int numberOfIndices)
-        {
-            var result = false;
-            numberOfIndices = 0;
-
-            fixed (byte* pSearchSpace = &searchSpace)
-            {
-                var searchStart = pSearchSpace;
-                var offset = 0;
-
-                while (true)
-                {
-                    if (Vector.IsHardwareAccelerated)
-                    {
-                        // Check Vector lengths
-                        if (length - Vector<byte>.Count >= offset)
-                        {
-                            Vector<byte> values = GetVector(value);
-                            do
-                            {
-                                var vFlaggedMatches = Vector.Equals(Unsafe.Read<Vector<byte>>(searchStart + offset), values);
-                                if (!vFlaggedMatches.Equals(Vector<byte>.Zero))
-                                {
-                                    // Found match, reuse Vector values to keep register pressure low
-                                    values = vFlaggedMatches;
-                                    break;
-                                }
-
-                                offset += Vector<byte>.Count;
-                            } while (length - Vector<byte>.Count >= offset);
-
-                            // Found match? Perform secondary search outside out of loop, so above loop body is small
-                            if (length - Vector<byte>.Count >= offset)
-                            {
-                                // Find offset of first match
-                                offset += LocateFirstFoundByte(values);
-                                // goto rather than inline return to keep function smaller
-                                goto exitFixed;
-                            }
-                        }
-                    }
-
-                    ulong flaggedMatches = 0;
-                    // Check ulong length
-                    while (length - sizeof(ulong) >= offset)
-                    {
-                        flaggedMatches = SetLowBitsForByteMatch(*(ulong*)(searchStart + offset), value);
-                        if (flaggedMatches != 0)
-                        {
-                            // Found match
-                            break;
-                        }
-
-                        offset += sizeof(ulong);
-                    }
-
-                    // Found match? Perform secondary search outside out of loop, so above loop body is small
-                    if (length - sizeof(ulong) >= offset)
-                    {
-                        // Find offset of first match
-                        offset += LocateFirstFoundByte(flaggedMatches);
-                        // goto rather than inline return to keep function smaller
-                        goto exitFixed;
-                    }
-
-                    // Haven't found match, scan through remaining
-                    for (; offset < length; offset++)
-                    {
-                        if (*(searchStart + offset) == value)
-                        {
-                            // goto rather than inline return to keep loop body small
-                            goto exitFixed;
-                        }
-                    }
-
-                    // No Matches
-                    result = true;
-                    break;
-
-                    exitFixed:;
-                    indices[numberOfIndices++] = offset++;
-                    if (numberOfIndices >= indices.Length)
-                        break;
-                }
-            }
-
-            return result && numberOfIndices < indices.Length;
-        }
-
-
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private static int LocateFirstFoundByte(Vector<byte> match)
         {
@@ -486,4 +264,3 @@ namespace System
         private const ulong filterByteHighBitsInUlong = (byteBroadcastToUlong >> 1) | (byteBroadcastToUlong << (sizeof(ulong) * 8 - 1));
     }
 }
-
