+++ tests/System.Buffers.Primitives.Tests/CastTests.cs
@@ -2,6 +2,7 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
+using System.Binary;
 using Xunit;
 
 namespace System.Buffers.Tests
