+++ tests/System.Buffers.Primitives.Tests/UsageScenarioTests.cs
@@ -2,6 +2,7 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using System.Binary;
 using System.Runtime.InteropServices;
 using Xunit;
 using Xunit.Abstractions;
