+++ tests/System.IO.Pipelines.Tests/ReadableBufferFacts.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
+using System.Binary;
 using System.Buffers;
 using System.Buffers.Pools;
 using System.Collections.Generic;
